Move AsyncStatus to constants & refactor types imports
