+++ src/index.ts
@@ -1,5 +1,6 @@
 export { default as action } from './action';
 export { default as asyncState, type AsyncState } from './async-state';
+export { AsyncStatus } from './constants';
 export { AwaitableEvent, flush, Registry } from './core';
 export { default as effect } from './effect';
 export { default as familyState } from './family-state';
@@ -16,15 +17,14 @@ export {
 export { default as scenario, type Config as ScenarioConfig } from './scenario';
 export { default as selector } from './selector';
 export { default as state, type State } from './state';
-export {
-  type AsyncSetter,
-  AsyncStatus,
-  type AsyncValue,
-  type FamilyState,
-  type InferReadableType,
-  type ReadableAsyncState,
-  type ReadableState,
-  type Setter,
-  type WritableAsyncState,
-  type WritableState,
+export type {
+  AsyncSetter,
+  AsyncValue,
+  FamilyState,
+  InferReadableType,
+  ReadableAsyncState,
+  ReadableState,
+  Setter,
+  WritableAsyncState,
+  WritableState,
 } from './types';
