+++ src/types/index.ts
@@ -1,5 +1,4 @@
 export { type default as AsyncSetter } from './AsyncSetter';
-export { default as AsyncStatus } from './AsyncStatus';
 export { type default as AsyncValue } from './AsyncValue';
 export { type default as FamilyState } from './FamilyState';
 export { type default as Id } from './Id';
